Add product price and product original price components
