+++ src/Sylius/Bundle/ShopBundle/Twig/Component/Common/ProductCardComponent.php
@@ -14,15 +14,12 @@ declare(strict_types=1);
 namespace Sylius\Bundle\ShopBundle\Twig\Component\Common;
 
 use Doctrine\Common\Collections\Collection;
-use Sylius\Bundle\MoneyBundle\Templating\Helper\FormatMoneyHelperInterface;
 use Sylius\Component\Channel\Context\ChannelContextInterface;
 use Sylius\Component\Core\Calculator\ProductVariantPricesCalculatorInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\ChannelPricingInterface;
 use Sylius\Component\Core\Model\ProductInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
-use Sylius\Component\Currency\Context\CurrencyContextInterface;
-use Sylius\Component\Locale\Context\LocaleContextInterface;
 use Sylius\Component\Product\Resolver\ProductVariantResolverInterface;
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 use Symfony\UX\TwigComponent\Attribute\ExposeInTemplate;
@@ -37,10 +34,7 @@ class ProductCardComponent
     public function __construct(
         private readonly ProductVariantResolverInterface $productVariantResolver,
         private readonly ChannelContextInterface $channelContext,
-        private readonly CurrencyContextInterface $currencyContext,
         private readonly ProductVariantPricesCalculatorInterface $productVariantPricesCalculator,
-        private readonly LocaleContextInterface $localeContext,
-        private readonly FormatMoneyHelperInterface $formatMoneyHelper,
     ) {
     }
 
@@ -57,24 +51,6 @@ class ProductCardComponent
         $this->variant = $variant;
 
         return $this->variant;
-    }
-
-    #[ExposeInTemplate(name: 'price')]
-    public function getPrice(): string
-    {
-        $price = $this->productVariantPricesCalculator
-            ->calculate($this->variant, ['channel' => $this->channelContext->getChannel()]);
-
-        return $this->formatPrice($price);
-    }
-
-    #[ExposeInTemplate(name: 'original_price')]
-    public function getOriginalPrice(): string
-    {
-        $price = $this->productVariantPricesCalculator
-            ->calculateOriginal($this->variant, ['channel' => $this->channelContext->getChannel()]);
-
-        return $this->formatPrice($price);
     }
 
     /** @return Collection<array-key, ChannelPricingInterface> */
@@ -96,14 +72,5 @@ class ProductCardComponent
             >
             $this->productVariantPricesCalculator
                 ->calculate($this->variant, ['channel' => $this->channelContext->getChannel()]);
-    }
-
-    private function formatPrice(int $price): string
-    {
-        return $this->formatMoneyHelper->formatAmount(
-            $price,
-            $this->currencyContext->getCurrencyCode(),
-            $this->localeContext->getLocaleCode(),
-        );
     }
 }
