+++ src/Sylius/Bundle/ShopBundle/templates/Product/component/original_price.html.twig
@@ -1,0 +1,1 @@
+{{ price }}
