+++ src/Sylius/Bundle/ShopBundle/templates/Product/component/price.html.twig
@@ -1,0 +1,1 @@
+{{ price }}
